Update server client-delete procedure: delete by host:port peer
The old version of client-delete procedure deletes given the client-name;
If two clients have the same name, if one of the clients request exit
the two clients will be deleted. YES. The another guy will be kicked!
No mercy.

To avoid that I add the client-id function to get the host:port of peer
based on the another function socket-id. So now the client-delete is
based on the client-id instead client-name.

As plus I keep all the functions separated by just one newline.
To avoid.

Added functions:
* socket-id :: socket -> host:port
* client-id :: client -> host:port