Use defparameter instead defconstant
This occurs because mainly that:

"SBCL compiles your file and then loads the result (the xxx-TMP.fasl
file), and for that particular implementation, the defconstant form is
thus evaluated twice in the same environment. A compiler needs not
actually evaluate the form during compilation (it could declare it
internally somehow so that further uses of the constants can be
inlined) but this is a valid compilation strategy. "

ref: https://stackoverflow.com/a/34801400/3749971