Fix build.lisp dynamic load by using uiop/os:getcwd
There is a semantic difference of a pathname and a string in Common
Lisp. sb-posix:getcwd only return a string and it doesn't works really
well with cffi:*foreign-library-directories*.

This fix everything! Now I can deploy the software with the dynamic
libraries in the root of the directory of lisp-chat binary.